+++ R/scale-continuous.r
@@ -74,7 +74,7 @@ NULL
 #'
 #' @export
 scale_x_continuous <- function(name = waiver(), breaks = waiver(),
-                               minor_breaks = waiver(), n_breaks = NULL,
+                               minor_breaks = waiver(), n.breaks = NULL,
                                labels = waiver(), limits = NULL,
                                expand = waiver(), oob = censor,
                                na.value = NA_real_, trans = "identity",
@@ -82,7 +82,7 @@ scale_x_continuous <- function(name = waiver(), breaks = waiver(),
                                sec.axis = waiver()) {
   sc <- continuous_scale(
     c("x", "xmin", "xmax", "xend", "xintercept", "xmin_final", "xmax_final", "xlower", "xmiddle", "xupper", "x0"),
-    "position_c", identity, name = name, breaks = breaks, n_breaks = n_breaks,
+    "position_c", identity, name = name, breaks = breaks, n.breaks = n.breaks,
     minor_breaks = minor_breaks, labels = labels, limits = limits,
     expand = expand, oob = oob, na.value = na.value, trans = trans,
     guide = guide, position = position, super = ScaleContinuousPosition
@@ -95,7 +95,7 @@ scale_x_continuous <- function(name = waiver(), breaks = waiver(),
 #' @rdname scale_continuous
 #' @export
 scale_y_continuous <- function(name = waiver(), breaks = waiver(),
-                               minor_breaks = waiver(), n_breaks = NULL,
+                               minor_breaks = waiver(), n.breaks = NULL,
                                labels = waiver(), limits = NULL,
                                expand = waiver(), oob = censor,
                                na.value = NA_real_, trans = "identity",
@@ -103,7 +103,7 @@ scale_y_continuous <- function(name = waiver(), breaks = waiver(),
                                sec.axis = waiver()) {
   sc <- continuous_scale(
     c("y", "ymin", "ymax", "yend", "yintercept", "ymin_final", "ymax_final", "lower", "middle", "upper", "y0"),
-    "position_c", identity, name = name, breaks = breaks, n_breaks = n_breaks,
+    "position_c", identity, name = name, breaks = breaks, n.breaks = n.breaks,
     minor_breaks = minor_breaks, labels = labels, limits = limits,
     expand = expand, oob = oob, na.value = na.value, trans = trans,
     guide = guide, position = position, super = ScaleContinuousPosition